requirements.txt
fastapi==0.96.0
uvicorn[standard]==0.22.0
loguru==0.7.0
transformers==4.26.1
torch
numpy==1.25.2
pandas==1.5.3
pydantic==1.10.14
underthesea==1.3.3
opentelemetry-api==1.19.0
opentelemetry-sdk==1.19.0
opentelemetry-instrumentation-asgi==0.40b0
opentelemetry-instrumentation-fastapi==0.40b0
opentelemetry-instrumentation-requests==0.40b0
opentelemetry-instrumentation-logging==0.40b0
opentelemetry-exporter-jaeger==1.19.0
opentelemetry-exporter-otlp-proto-grpc==1.19.0
opentelemetry-exporter-prometheus==1.12.0rc1
python-multipart==0.0.6
prometheus-client==0.17.1